+++ src/main/java/org/omoknoone/ppm/domain/projectDashboard/service/ProjectDashboardService.java
@@ -3,6 +3,7 @@ package org.omoknoone.ppm.domain.projectDashboard.service;
 import java.util.List;
 
 import org.omoknoone.ppm.domain.projectDashboard.aggregate.ProjectDashboard;
+import org.omoknoone.ppm.domain.projectDashboard.dto.ProjectDashboardDTO;
 
 public interface ProjectDashboardService {
 	List<ProjectDashboardDTO> viewProjectDashboardByProjectId(String projectId);
