+++ src/main/java/org/omoknoone/ppm/domain/requirements/service/RequirementsService.java
@@ -1,12 +1,8 @@
 package org.omoknoone.ppm.domain.requirements.service;
 
-import java.util.List;
-
-import org.omoknoone.ppm.domain.requirements.dto.ProjectRequirementsListDTO;
 import org.omoknoone.ppm.domain.requirements.dto.RequirementsDTO;
 import org.omoknoone.ppm.domain.requirements.vo.RequestModifyRequirement;
 import org.omoknoone.ppm.domain.requirements.vo.ResponseRequirement;
-import org.springframework.transaction.annotation.Transactional;
 
 public interface RequirementsService {
